Fix handling of manufacturer data in Bluetooth advertisement parsing
There were two bugs in the handling of manufacturer data
when parsing advertisement packages:
First, not every call to onBluetoothLEAdvertisementReceived
contains manufacturer data. In the call where no manufacturer data
is given, data given earlier should not be deleted.
Second, manufacturer data was only processed when also data
about the services was present, which is sometimes not the case.
This patch fixes both bugs.

Pick-to: 6.2
Change-Id: I3d0798e0a19dd7ade6476e2f9b37397094aebf60